--- src/main/java/calculator/operator/DivideOperator.java
@@ -1,8 +1,0 @@
-package calculator.operator;
-
-public class DivideOperator implements Operator {
-  @Override
-  public int calculate(int left, int right) {
-    return left / right;
-  }
-}
