+++ src/main/java/calculator/type/Operator.java
@@ -1,4 +1,4 @@
-package calculator.operator;
+package calculator.type;
 
 public interface Operator {
   int calculate(int left, int right);
